pyproject.toml
[project]
name = "learn-language"
version = "0.1.0"
description = ""
authors = [
    {name = "vitalyionov",email = "vitalyionov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "authlib (>=1.6.0,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

